requirements.txt
Flask>=1.1.1
pyyaml>=5.3
